+++ sdk/metric/exemplar.go
@@ -1,22 +1,12 @@
 // Copyright The OpenTelemetry Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+// SPDX-License-Identifier: Apache-2.0
 
 package metric // import "go.opentelemetry.io/otel/sdk/metric"
 
 import (
 	"os"
 	"runtime"
+	"slices"
 
 	"go.opentelemetry.io/otel/sdk/metric/internal/exemplar"
 	"go.opentelemetry.io/otel/sdk/metric/internal/x"
@@ -40,8 +30,7 @@ func reservoirFunc[N int64 | float64](agg Aggregation) func() exemplar.Reservoir
 		// use AlignedHistogramBucketExemplarReservoir.
 		a, ok := agg.(AggregationExplicitBucketHistogram)
 		if ok && len(a.Boundaries) > 0 {
-			cp := make([]float64, len(a.Boundaries))
-			copy(cp, a.Boundaries)
+			cp := slices.Clone(a.Boundaries)
 			return func() exemplar.Reservoir[N] {
 				bounds := cp
 				return exemplar.Histogram[N](bounds)
@@ -63,7 +52,7 @@ func reservoirFunc[N int64 | float64](agg Aggregation) func() exemplar.Reservoir
 			//   This Exemplar reservoir MAY take a configuration parameter for
 			//   the size of the reservoir. If no size configuration is
 			//   provided, the default size MAY be the number of possible
-			//   concurrent threads (e.g. numer of CPUs) to help reduce
+			//   concurrent threads (e.g. number of CPUs) to help reduce
 			//   contention. Otherwise, a default size of 1 SHOULD be used.
 			n = runtime.NumCPU()
 			if n < 1 {