+++ sdk/metric/internal/exemplar/reservoir_test.go
@@ -1,16 +1,5 @@
 // Copyright The OpenTelemetry Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+// SPDX-License-Identifier: Apache-2.0
 
 package exemplar
 
@@ -125,7 +114,7 @@ func ReservoirTest[N int64 | float64](f factory[N]) func(*testing.T) {
 			r.Collect(&dest)
 			assert.Len(t, dest, n, "multiple offers did not fill reservoir")
 
-			// Ensure the collect reset also resets any couting state.
+			// Ensure the collect reset also resets any counting state.
 			for i := 0; i < n+1; i++ {
 				v := N(i)
 				r.Offer(ctx, staticTime, v, nil)